More clean up and updates, Also got dashboard front-end again
Dashboard, now shows the New Idea Box (Form) when clicking on the Add
Idea icon.

Fixed layout issues and bootstrap which I had broken yesterday while
updating.

Fixed typo in the header of the page for the viewport meta def.

Notification box is functioning again.

Fake comments are also working again, now it's just a thing of making
it unfaked.